The timeout value is now stored in the XML-configuration. Since it is only used by objects which implement the Drawable interface, it is marked as not required. For the same reason "isVisible" is marked as not required, too

+++ src/de/uniluebeck/itm/spyglass/plugin/Drawable.java
@@ -1,9 +1,10 @@
 /*
- * -------------------------------------------------------------------------------- This file is
- * part of the WSN visualization framework SpyGlass. Copyright (C) 2004-2007 by the SwarmNet
- * (www.swarmnet.de) project SpyGlass is free software; you can redistribute it and/or modify it
- * under the terms of the BSD License. Refer to spyglass-licence.txt file in the root of the
- * SpyGlass source tree for further details.
+ * --------------------------------------------------------------------------------
+ * This file is part of the WSN visualization framework SpyGlass. Copyright (C)
+ * 2004-2007 by the SwarmNet (www.swarmnet.de) project SpyGlass is free
+ * software; you can redistribute it and/or modify it under the terms of the BSD
+ * License. Refer to spyglass-licence.txt file in the root of the SpyGlass
+ * source tree for further details.
  * --------------------------------------------------------------------------------
  */
 package de.uniluebeck.itm.spyglass.plugin;
@@ -14,8 +15,9 @@ import de.uniluebeck.itm.spyglass.drawing.DrawingObject;
 import de.uniluebeck.itm.spyglass.gui.view.DrawingArea;
 
 /**
- * Plugins which implement this interface want to draw on the drawingArea. They will be called on
- * redrawEvents to return their current List of drawingObjects via the method defined below.
+ * Plugins which implement this interface want to draw on the drawingArea. They
+ * will be called on redrawEvents to return their current List of drawingObjects
+ * via the method defined below.
  * 
  * 
  * @author dariush
@@ -26,7 +28,7 @@ public interface Drawable {
 	/**
 	 * TODO: really needed??
 	 */
-	public float getTimeout();
+	public int getTimeout();
 	
 	/**
 	 * Return if this plugin is currently visible.
